fix: resolve 55 HIGH severity bugs across backend and frontend
Backend:
- channel-manager: atomic redis incr+expire via multi/exec (race condition)
- conversation-state: atomic jsonb append via onConflictDoUpdate (race condition)
- conversations.messaging: onConflictDoNothing for findOrCreateConversation (race condition)
- auth.routes: remove token console.log, fix 23505 error code, narrow reset-password query to 2h window
- auth.service: fallback slug when tenantName produces empty string
- billing.service: fix grace period logic to use currentPeriodEnd + 3 days
- error-handler: sanitize Zod field paths in production
- channels.service: atomic upsert via onConflictDoUpdate + unique index
- llm-client: add 15s timeout to OpenAI calls
- evolution-api.client: add 10s fetch timeout
- wompi.webhook: return 404 for missing payments
- reminder.job: mark reminderSent before sending to prevent retry storm
- billing.service: eliminate N+1 query on expired subscriptions
- campaign-sender.job: checkpoint progress every 10 messages
- orders-mcp-server: allow free products (price===0), limit scans to 50, validate stock
- appointments-mcp-server: reject past dates, disambiguate cancelAppointment
- payments-mcp-server: enforce integer amounts for COP
- catalog-mcp-server: implement categoryName filter, limit scans to 50
- mcp-client: robust JSON extractor with brace-depth counting
- scheduling.engine: document hardcoded business hours limitation
- conversations.messaging: normalize phone numbers before lookup

Frontend:
- inbox/page: fix SSE reconnect leak by removing loadConversations from effect deps
- channels/page: add toast.error in refreshQR catch block
- kanban/page: add toast.error in NewColumnModal submit catch block
- appointments, customers, orders, settings, superadmin: various UX/stability improvements
- Modal: add focus trap, Escape handler, aria attributes

DB Schema:
- channels.ts: add unique index on channel_sessions(tenantId, channel)

+++ apps/api/src/mcp/servers/payments-mcp-server.ts
@@ -17,7 +17,7 @@ export const paymentsMCPServer: MCPServer = {
       description: 'Genera un link de pago Wompi para un pedido existente o monto específico',
       parameters: z.object({
         orderNumber: z.string().optional().describe('Número de pedido para pagar'),
-        amount: z.number().positive().optional().describe('Monto a pagar (si no hay pedido)'),
+        amount: z.number().positive().int().optional().describe('Monto a pagar (si no hay pedido)'),
         description: z.string().optional().describe('Descripción del pago'),
       }),
       execute: async (params, ctx) => {
